Adds +1 to getMonth()
